Add RepoCloud Deployment Option for Formbricks
This commit introduces a new deployment option for Formbricks, enabling users to deploy on RepoCloud with just one click. By adding the "Deploy on RepoCloud" button alongside the existing Railway option, we offer more flexibility and choices for the community in hosting their applications. This update makes it easier for users to take advantage of RepoCloud's efficient and cost-effective cloud hosting services for open source applications.